Eliminate duplication between QuotaUpdatesMonitor and UsageStatsMonitor
